fixing flatMap?
