+++ build.sbt
@@ -4,7 +4,7 @@ enablePlugins(ScalaJSBundlerPlugin)
 
 libraryDependencies ++= Seq(
   "org.scala-js" %%% "scalajs-dom" % "0.9.6", // This has no runtime cost. We only use it for `Debug.log` // @TODO[Elegance] Reconsider
-  "org.scalatest" %%% "scalatest" % "3.0.5" % Test
+  "org.scalatest" %%% "scalatest" % "3.0.6" % Test
 )
 
 useYarn := true
